% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer.R
\name{as.days}
\alias{as.days}
\title{Reduce Timestamp to Days}
\usage{
as.days(x)
}
\arguments{
\item{x}{object which can be converted to Date.}
}
\value{
Returns an object of same length as argument \code{x} reduced
to day only.
}
\description{
Small utility function which converts a timestamp/POSIXt to
day only, accounting for time zone if set.
}
\author{
Reto
}
