% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer.R
\name{load_logs}
\alias{load_logs}
\title{Loading Logs From Database}
\usage{
load_logs(
  con,
  ips = NULL,
  start = NULL,
  end = NULL,
  limit = NULL,
  quiet = FALSE
)
}
\arguments{
\item{con}{object of class \code{SQLiteConnection}}

\item{ips}{\code{NULL} (loads logs for all IPs) or a character
vector with valid IPs to restrict the logs loaded
to specific IP addresses.}

\item{start}{either \code{NULL} or POSIXt, defines the beginning
of the logs to be returned if set.}

\item{end}{either \code{NULL} or POSIXt, defines the beginning
of the logs to be returned if set.}

\item{limit}{\code{NULL} (default) or positive integer, will limit the
number of logs (useful for testing).}

\item{quiet}{Logical, defaults to \code{FALSE}.}
}
\value{
Returns an object of class \code{apachelogs}, a list of length two with
two data frames containing the access logs as well as the error logs.
This object can be handed over to the \code{\link[=analyze_logs]{analyze_logs()}} function to retrieve
the log statistics.
}
\description{
After the raw log files have been parsed an put into
a sqlite3 database via \code{\link[=parse_file]{parse_file()}} this function
can be called to read the data from that database
and prepare the data for further analysis.
}
\details{
The arguments \code{start} and \code{end} allow to limit the logs
loaded to a specific period. For convenience \verb{start >=} and \verb{end <}
is used, so one can retrieve logs for one specific day by supplying
e.g., \code{start = "2026-01-03"} and \code{end = "2026-01-04"} which will
include all logs for 2026-01-03.
}
\author{
Reto
}
