% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_file}
\alias{parse_file}
\title{Parsing Logfile}
\usage{
parse_file(
  file,
  con = NULL,
  n = 10L,
  type = NULL,
  verbose = FALSE,
  maxbatches = .Machine$integer.max,
  warn = TRUE
)
}
\arguments{
\item{file}{character, path to the logfile to be parsed.}

\item{con}{either \code{NULL} (just returns a data.frame) or
an object of class \code{SQLiteConnection}. In this case
the data are not returned but stored in the SQLite database.}

\item{n}{positive numeric, number of lines to parse in one batch
from the logfile and written to the database.}

\item{type}{either \code{NULL} (auto-detect) or one of \code{"error"} (when parsing
error logfiles) or \code{"access"} (if parsing access logs). If set
\code{NULL} it will be derived from the file name (if the file name
contains \code{"error"} or \code{"access"}) or stops.}

\item{verbose}{If set \code{TRUE}, some output will be printed.}

\item{maxbatches}{Maximum number of batches of \code{n} rows to be read.}

\item{warn}{Logical, defaults to \code{TRUE}. Can be suppressed by setting
it \code{FALSE}.}
}
\value{
Invisibly returns the number of parsed lines.
}
\description{
Parsing Logfile
}
\examples{
\dontrun{
  FILE   <- "my_access_log"
  con    <- open_database("test.sqlite3")
  nlines <- parse_file(FILE, con = con, n = 1e5, verbose = TRUE)
  msg    <- dbGetQuery(con, "SELECT * FROM messages")
  log    <- dbGetQuery(con, "SELECT * FROM logs")
  dbDisconnect(con)
}

}
