% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer.R
\name{analyze_logs}
\alias{analyze_logs}
\title{Analyzing Apache Logs}
\usage{
analyze_logs(x, unit = c("hours", "days", "minutes", "seconds"), ...)
}
\arguments{
\item{x}{object of class apachelogs as returned by the function
\code{\link[=load_logs]{load_logs()}}.}

\item{unit}{character, defines on which temporal level the data
are aggregated.}
}
\description{
Analyzing Apache Logs
}
\author{
Reto
}
