% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{open_database}
\alias{open_database}
\title{Open (or create) SQLite Database}
\usage{
open_database(dbfile)
}
\arguments{
\item{file}{path to the SQLite database.}
}
\value{
Returns SQLite database connection.
}
\description{
Open (or create) SQLite Database
}
\author{
Reto
}
