% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_file}
\alias{parse_file}
\title{Parsing Apache2 Logfile}
\usage{
parse_file(
  file,
  con = NULL,
  n = 10L,
  type = NULL,
  verbose = FALSE,
  maxbatches = .Machine$integer.max,
  warn = TRUE
)
}
\arguments{
\item{file}{character, path to the logfile to be parsed.}

\item{con}{either \code{NULL} (just returns a data.frame) or
an object of class \code{SQLiteConnection}. In this case
the data are not returned but stored in the SQLite database.}

\item{n}{positive numeric, number of lines to parse in one batch
from the logfile and written to the database.}

\item{type}{either \code{NULL} (auto-detect) or one of \code{"error"} (when parsing
error logfiles) or \code{"access"} (if parsing access logs). If set
\code{NULL} it will be derived from the file name (if the file name
contains \code{"error"} or \code{"access"}) or stops.}

\item{verbose}{If set \code{TRUE}, some output will be printed.}

\item{maxbatches}{Maximum number of batches of \code{n} rows to be read.}

\item{warn}{Logical, defaults to \code{TRUE}. Can be suppressed by setting
it \code{FALSE}.}
}
\value{
Invisibly returns the number of parsed lines.
}
\description{
Parsing Apache2 Logfile
}
\details{
Will parse both access logs and error logs. Logfiles
can contain incomplete or broken lines. Thus, we will remove
all lines which fail parsing, and also remove all files where
we can't properly decode the time information.
}
\examples{
library("apache2logparser")

dir <- system.file("extdata", package = "apache2logparser")
stopifnot(file.exists(FILE_A <- file.path(dir, "www.example.com_access_log")))
stopifnot(file.exists(FILE_E <- file.path(dir, "www.example.com_error_log")))

# Create random file name for sqlite3 database.
DBFILE <- basename(tempfile(fileext = "_demo.sqlite3"))

# Connecting to database
con <- open_database(DBFILE)

# Parsing access log file and error log file; both directly
# write to database via the database connection
an   <- parse_file(FILE_A, con = con, n = 30, verbose = TRUE)
en   <- parse_file(FILE_E, con = con, n = 30, verbose = TRUE)

# Reading all data from the three different tables created
msg  <- dbGetQuery(con, "SELECT * FROM messages")
alog <- dbGetQuery(con, "SELECT * FROM access_logs")
elog <- dbGetQuery(con, "SELECT * FROM error_logs")

# Disconnecting from database
dbDisconnect(con)

# Some output for testing
message("Size of table 'messages':    ", paste(dim(msg), collapse = " x "))
message("Size of table 'access_logs': ", paste(dim(alog), collapse = " x "))
message("Size of table 'error_logs':  ", paste(dim(elog), collapse = " x "))

# Removing demo file
file.remove(DBFILE)

}
\author{
Reto
}
