% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_file}
\alias{parse_file}
\title{Parsing Logfile}
\usage{
parse_file(file, n = 10L, type = NULL, ...)
}
\arguments{
\item{file}{character, path to the logfile to be parsed.}

\item{n}{positive numeric, number of lines to parse in one batch.}

\item{type}{either \code{NULL} (auto-detect) or one of \code{"error"} (when parsing
error logfiles) or \code{"access"} (if parsing access logs). If set
\code{NULL} it will be derived from the file name (if the file name
contains \code{"error"} or \code{"access"}) or stops.}
}
\description{
Parsing Logfile
}
